+++ src/LEDUtils.cpp
@@ -6,7 +6,7 @@ breath_compute (BreathState *state) {
     // This code is adapted from FastLED lib8tion.h as of dd5d96c6b289cb6b4b891748a4aeef3ddceaf0e6
     // Eventually, we should consider just using FastLED
 
-    uint8_t i = millis()/12; 
+    uint8_t i = (uint16_t)millis()/12; 
 
     if( i & 0x80) {
         i = 255 - i;
